refactor: compose Exceptions based on entity
