+++ src/main/kotlin/me/insiro/home/server/post/exception/category/CategoryConflictException.kt
@@ -1,4 +1,4 @@
-package me.insiro.home.server.post.exception
+package me.insiro.home.server.post.exception.category
 
 import me.insiro.home.server.application.exception.AbsException
 import me.insiro.home.server.post.entity.Category
