+++ src/main/kotlin/me/insiro/home/server/post/exception/post/PostNotFoundException.kt
@@ -1,4 +1,4 @@
-package me.insiro.home.server.post.exception
+package me.insiro.home.server.post.exception.post
 
 import me.insiro.home.server.application.exception.AbsException
 import me.insiro.home.server.post.entity.Post
